cambios en el backend (Al crear, y registrar usuarios y cursos  no los habilitaba) y front (Mostrar solo cursos que esten habilitados)

+++ frontend/gestionUsuarios/src/app/Component/curso-list/curso-list.component.ts
@@ -42,10 +42,13 @@ export class CursoListComponent implements OnInit{
     this.idAdmin = Number(localStorage.getItem('idAdmin'))
   }
 
-  deleteCurso(id: number): void{
-    this.cursoService.deleteCurso(id).subscribe(() => {
-      this.cursos = this.cursos.filter(c => c.id !== id);
-    })
+  deleteCurso(id: number, nombreCurso: string): void{
+    const ok = confirm("¿Estas seguro que deseas eliminar el curso "+nombreCurso.toUpperCase()+" ?")
+    if(ok){
+      this.cursoService.deleteCurso(id).subscribe(() => {
+        this.cursos = this.cursos.filter(c => c.id !== id);
+      })  
+    }
   }
 
   btnUpdateCurso(isUpdate: boolean, id: number): void{
